Debugging

+++ utils.py
@@ -1,7 +1,5 @@
 import warnings
 
-import numpy as np
-
 
 def index_to_position(index):
     """ Transform 1d int index to 2d tuple position """
@@ -36,7 +34,7 @@ def position_to_index(position):
 
 
 def grid_to_string(grid):
-    grid_flatten = list(grid.flatten())
+    grid_flatten = list(grid.astype(int).flatten())
     grid_flatten_converted = [str(el+1).replace('0','O').replace('1','-').replace('2','X') for el in grid_flatten]
     grid_hash = "".join(grid_flatten_converted)
     return grid_hash
